+++ shaderflow/__main__.py
@@ -1,12 +1,13 @@
 import sys
 
+from parsenaut._cyclopts import Launcher
+
 import shaderflow
-from shaderflow.launcher import SceneLauncher
 
 
 def main():
-    app = SceneLauncher()
-    app.common(package=shaderflow.package)
+    app = Launcher(keyword="Scene")
+    app.smart(package=shaderflow.package)
     app.cli(sys.argv[1:])
 
 if __name__ == "__main__":